LayoutTests/storage/indexeddb/modern/transaction-scheduler-5-expected.txt

This test creates two object stores. It then opens two long running read-only transactions, one per object store. It then opens a read-write transaction to both object stores. The read-only transactions both need to finish before the read-write transaction starts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Upgrade needed: Old version - 0 New version - 1
ALERT: versionchange transaction completed
ALERT: Success opening database connection - Starting transactions
ALERT: Starting write transaction
ALERT: Read transaction complete - [object Event]
ALERT: Read transaction complete - [object Event]
ALERT: Write to OS1 successful
ALERT: Write to OS2 successful
PASS successfullyParsed is true

TEST COMPLETE
